requirements.txt
# BaiMuras Platform Dependencies - Updated for Security
# Core Flask framework and extensions
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-CORS==6.0.1
Flask-Limiter==3.12
Flask-Talisman==1.1.0

# Database
SQLAlchemy==2.0.41
Alembic==1.16.2

# Authentication and Security
PyJWT==2.10.1
Werkzeug==3.1.3
bcrypt==4.3.0

# Forms and Validation
WTForms==3.2.1

# Email
Flask-Mail==0.10.0

# HTTP Requests
requests==2.32.4

# Task Queue (Celery)
celery==5.5.3
redis==6.2.0

# Environment Variables
python-dotenv==1.1.0

# Development and Testing
pytest==8.4.1
pytest-flask==1.3.0

# Code Quality
autopep8==2.3.2
flake8==7.3.0

# Production Server
gunicorn==23.0.0

# Utilities
python-dateutil==2.9.0.post0

# System monitoring
psutil==7.0.0

# Additional dependencies
blinker==1.9.0
click==8.2.1
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
typing-extensions==4.14.0
